<fix> create a file which exports repeated argTypes
