+++ src/stories/Footer.stories.tsx
@@ -1,99 +1,7 @@
 import { Story } from '@storybook/react';
 import { Footer } from '../lib-components/Footer';
 import './storiesGlobals.css';
-
-export const footerArgTypes = {
-  footerTitle: {
-    name: 'title',
-    type: { name: 'string', required: false },
-    description: "String wich defines footer's title",
-    control: { type: 'text' },
-    table: {
-      category: 'Footer'
-    }
-  },
-  footerTelephone: {
-    name: 'telephone',
-    type: { name: 'string', required: false },
-    description: "String wich defines footer's telephone",
-    control: { type: 'text' },
-    table: {
-      category: 'Footer'
-    }
-  },
-  footerTelephoneComplement: {
-    name: 'telephone complement',
-    type: { name: 'string', required: false },
-    description: "String wich defines footer's telephone complement",
-    control: { type: 'text' },
-    table: {
-      category: 'Footer'
-    }
-  },
-  footerEmail: {
-    name: 'email',
-    type: { name: 'string', required: false },
-    description: "String wich defines footer's email",
-    control: { type: 'text' },
-    table: {
-      category: 'Footer'
-    }
-  },
-  footerLink: {
-    name: 'link',
-    type: { name: 'string', required: false },
-    description: "String wich defines footer's link",
-    control: { type: 'text' },
-    table: {
-      category: 'Footer'
-    }
-  },
-  footerTextLink: {
-    name: 'text link',
-    type: { name: 'string', required: false },
-    description: "String wich defines footer's text link",
-    control: { type: 'text' },
-    table: {
-      category: 'Footer'
-    }
-  },
-  footerDescription: {
-    name: 'description',
-    type: { name: 'string', required: false },
-    description: "String wich defines footer's description",
-    control: { type: 'text' },
-    table: {
-      category: 'Footer'
-    }
-  },
-  footerCopyrightText: {
-    name: 'copyright text',
-    type: { name: 'string', required: false },
-    description: "String wich defines footer's copyright text",
-    control: { type: 'text' },
-    table: {
-      category: 'Footer'
-    }
-  },
-  footerSignatureText: {
-    name: 'signature text',
-    type: { name: 'string', required: false },
-    description: "String wich defines footer's signature text",
-    control: { type: 'text' },
-    table: {
-      category: 'Footer'
-    }
-  },
-  footerSignatureLink: {
-    name: 'signature link',
-    type: { name: 'string', required: false },
-    description: "String wich defines footer's signature link",
-    control: { type: 'text' },
-    table: {
-      category: 'Footer'
-    }
-  }
-};
+import { footerArgTypes } from '../utils/argTypes';
 
 export default {
   title: 'Components/Footer',
@@ -102,7 +10,7 @@ export default {
       page: null
     }
   },
-  argTypes: { footerArgTypes }
+  argTypes: { ...footerArgTypes }
 };
 
 export interface FooterStoryProps {
